[Sehyun] - added: 로그인, 회원가입, API
